+++ CotEditor/Sources/CEEditorWrapper.m
@@ -121,12 +121,13 @@
     [[self splitViewController] addSubviewForViewController:editorViewController relativeTo:nil];
     [self setupEditorViewController:editorViewController baseView:nil];
     
+    // focus text view
+    [[self window] makeFirstResponder:[editorViewController textView]];
+    
     // TODO: Refactoring
     // -> This is probably not the best position to apply sytnax style to the text view.
     //    However as a quick fix, I put it here tentatively. It works. But should be refactored later. (2016-01 1024jp)
     [editorViewController applySyntax:[self syntaxStyle]];
-    
-    [self setFocusedTextView:[editorViewController textView]];
     
     [[NSNotificationCenter defaultCenter] addObserver:self
                                              selector:@selector(didChangeSyntaxStyle:)